@sfc-gh-aalam 2023-04-17
```
-- with changes
snowflake-connector-python-py38 ❯  hyperfine 'python -c "import snowflake.connector"'
Benchmark 1: python -c "import snowflake.connector"
  Time (mean ± σ):     651.6 ms ±   9.6 ms    [User: 1009.8 ms, System: 2065.5 ms]
  Range (min … max):   640.9 ms … 668.3 ms    10 runs

-- without changes
snowflake-connector-python-py38 ❯  hyperfine 'python -c "import snowflake.connector"'
Benchmark 1: python -c "import snowflake.connector"
  Time (mean ± σ):     755.8 ms ±  57.4 ms    [User: 1096.9 ms, System: 1953.0 ms]
  Range (min … max):   707.3 ms … 897.7 ms    10 runs
```